requirements.txt
PyPDF4
re
pdfminer.six
